1. 新增了脚本节点之间连接时，入参类型伴随来源节点的返回类型改变而改变。 2. 重新优化了Script项目脚本生成代码的缩进排版 3. 修复了Script中对于Double字面量错误的使用了Float解析的bug

+++ Serein.Script/SereinScriptParser.cs
@@ -1087,7 +1087,7 @@ namespace Serein.Script
             }
             else if (_currentToken.Type == TokenType.NumberDouble)
             {
-                var value = float.Parse(_currentToken.Value);
+                var value = double.Parse(_currentToken.Value);
                 NextToken();  // 消耗 double 浮点数
                 return new NumberDoubleNode(value).SetTokenInfo(factorToken);
             }
